+++ accommodation-catalog-service/src/main/java/com/onesquad/accommodation/AccommodationRepository.java
@@ -2,6 +2,6 @@ package com.onesquad.accommodation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccommodationRepository extends JpaRepository<Accommodation, Integer> {
+public interface AccommodationRepository extends JpaRepository<Accommodation, Long> {
 
 }
